+++ resources/views/livewire/profile-warga.blade.php
@@ -22,11 +22,13 @@
     <div class="container-fluid">
       <div class="card">
         <div class="card-header">
-          <div class="row">
-            <div class="col-12 text-right">
-              <a href="{{ route('edit.warga',$citizen->id) }}" class="btn btn-warning">Ubah Data</a>
+          @if (auth()->user()->role == 'admin-yayasan')
+            <div class="row">
+              <div class="col-12 text-right">
+                <a href="{{ route('edit.warga',$citizen->id) }}" class="btn btn-warning">Ubah Data</a>
+              </div>
             </div>
-          </div>
+          @endif
         </div>
         <div class="card-body">
           <div class="row">
@@ -108,9 +110,11 @@
             <div class="col-sm-12 col-md-6 col-lg-4 col-xl-4">
               <input type="text" wire:model="search" class="form-control rounded-pill" placeholder="Cari Nama Berkas">
             </div>
-            <div class="col-sm-12 col-md-6 col-lg-8 col-xl-8 text-right">
-              <button class="btn btn-success" data-toggle="modal" data-target="#unggahBerkas">Unggah Berkas</button>
-            </div>
+            @if (auth()->user()->role == 'admin-yayasan')
+              <div class="col-sm-12 col-md-6 col-lg-8 col-xl-8 text-right">
+                <button class="btn btn-success" data-toggle="modal" data-target="#unggahBerkas">Unggah Berkas</button>
+              </div>
+            @endif
           </div>
         </div>
         <div class="card-body">
@@ -140,7 +144,9 @@
                     </td>
                     <td style="height: 50px">
                       <button wire:click="download('{{ $document->file }}','{{ $document->nama_dokumen }}')" class="btn btn-info btn-sm">Unduh Berkas</button>
-                      <button wire:click="deleteConfirmation('{{ $document->file }}','{{ $document->id }}')" class="btn btn-danger btn-sm">Hapus Berkas</button>
+                      @if (auth()->user()->role == 'admin-yayasan')
+                        <button wire:click="deleteConfirmation('{{ $document->file }}','{{ $document->id }}')" class="btn btn-danger btn-sm">Hapus Berkas</button>
+                      @endif
                     </td>
                   </tr>
                   @endforeach
